// src/api/generated/models/CreatePengajuanBodySchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CreatePengajuanBodySchema = {
  propertyId: string;
  title: string;
  startDate: string;
  endDate: string;
  alasan?: string;
  jenisPeminjaman: 'eksklusif' | 'non-eksklusif';
};